add game routing
