+++ src/FE6/__init__.py
@@ -1,0 +1,2 @@
+import FE6.get_characters
+import FE6.get_growths
