+++ src/FE7/__init__.py
@@ -1,0 +1,2 @@
+import FE7.get_characters
+import FE7.get_growths
