+++ src/FE8/__init__.py
@@ -1,0 +1,2 @@
+import FE8.get_characters
+import FE8.get_growths
